+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-* Added option "Position of minimap buttons".
-* Added option "Expand to".
-* Fixed a bug where the panel would appear behind the screen.
+* Updated grab buttons system.